# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "De'Corrinna API",
        "description": "REST API backend De'Corrinna (Laravel 12 + PostgreSQL). Autentikasi memakai Bearer token (Laravel Sanctum).",
        "contact": {
            "name": "Tim Azatech",
            "email": "dev@azatech.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (dev)"
        }
    ],
    "paths": {
...
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Daftar kategori (villa)",
                "description": "Terpaginasi. Menyertakan `items_count` untuk kolom \"Jumlah Item\" di layar B4.",
                "operationId": "8337773a2c74418316bc7d900b21bfd5",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Cari berdasarkan nama.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Aktif",
                                "Nonaktif"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Category"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak punya permission categories:index",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Tambah kategori",
                "description": "`slug` boleh dikosongkan — otomatis diturunkan dari `name` dan dibuat unik per tenant.",
                "operationId": "75b0f0a6bf53d317ac1c98222a694749",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/{category}": {
            "put": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Ubah kategori",
                "operationId": "decc152d00e60f86768977bbbe517a13",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan (atau milik tenant lain)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Hapus kategori",
                "description": "Soft delete. Ditolak bila masih punya item — hapus/pindahkan itemnya lebih dulu.",
                "operationId": "59f9e8d62f6df5da4b508311692e8e94",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Terhapus"
                    },
                    "422": {
                        "description": "Masih punya item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "description": "Kategori = \"villa\" di katalog publik, sekaligus induk item di admin.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Villa De Corrinna"
                    },
                    "slug": {
                        "description": "Dipakai /api/villas/{slug}. Unik per tenant.",
                        "type": "string",
                        "example": "villa-de-corrinna"
                    },
                    "icon": {
                        "description": "Nama ikon lucide.",
                        "type": "string",
                        "example": "house",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Rumah utama kami — sanctuary yang tenang di antara pepohonan pinus.",
                        "nullable": true
                    },
                    "tagline": {
                        "type": "string",
                        "example": "Rumah utama dengan taman luas dan kolam privat.",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "Kawasan Gerbang Gunungsari, Pamijahan, Bogor",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "items_count": {
                        "description": "Kolom \"Jumlah Item\" di B4 — hanya ada pada endpoint list.",
                        "type": "integer",
                        "example": 12
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Facility"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CategoryInput": {
                "title": "CategoryInput",
                "description": "Body untuk POST/PUT kategori. Cerminan StoreCategoryRequest/UpdateCategoryRequest.",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Villa De Corrinna"
                    },
                    "slug": {
                        "description": "Opsional — diturunkan dari `name` bila kosong, lalu dibuat unik per tenant.",
                        "type": "string",
                        "example": "villa-de-corrinna"
                    },
                    "icon": {
                        "type": "string",
                        "example": "house",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "tagline": {
                        "type": "string",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "facility_ids": {
                        "description": "Sync fasilitas villa. Bila field ini tidak dikirim sama sekali, tautan lama dibiarkan apa adanya.",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Envelope response error standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan."
                    },
                    "errors": {
                        "description": "Peta field -> daftar pesan (hanya untuk error validasi).",
                        "type": "object",
                        "example": {
                            "email": [
                                "Email wajib diisi."
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Facility": {
                "title": "Facility",
                "description": "Cerminan FacilityResource.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Private Pool"
                    },
                    "icon": {
                        "description": "Nama ikon lucide.",
                        "type": "string",
                        "example": "waves",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Meta pagination standar untuk response terpaginasi.",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Envelope response sukses standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token pribadi dari Laravel Sanctum. Kirim sebagai header: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}